app/requirements.txt
six==1.17.0
cassandra-driver==3.29.2
click==8.1.8
colorama==0.4.6 ; sys_platform == 'win32'
geomet==0.2.1.post1
py4j==0.10.9.7
pyspark==3.5.5
venv-pack
pyarrow
fastparquet
pathvalidate==3.2.1
tqdm
